refactor(estimates): split the 1153-line detail page
estimates/[id]/page.tsx 1153 -> ~290 lines. Pure extraction, no behavior
change (typecheck + lint + build pass):

- detail-data.ts — row types + loadEstimateDetail() (estimate, line items,
  options, scheduled-visit count, deposit/final invoices, change orders).
- format.ts — STATUS_LABELS + formatDollars.
- sections/EstimateBanners.tsx — approved + expired banners and billing summary.
- sections/EstimateSummaryCard.tsx — totals, scope, room breakdown, internal
  margin, materials plan, pricing guardrails.
- sections/EstimateLineItems.tsx — multi-option / mobile / desktop layouts.
- sections/ApprovedHandoff.tsx — materials -> schedule -> final billing.

page.tsx keeps the header, status stepper, and the already-componentized
action panels (edit form, review, send, transition, internal notes, change
orders, danger zone).
